moved Android-specific stuff into androis.rs
